core.agents/interview_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pydantic>=2.0.0
typing-extensions>=4.0.0
asyncio>=3.4.3
aiohttp>=3.8.0
